product_catalog/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
Django==5.1
djangorestframework==3.15.2
idna==3.7
msgpack==1.0.8
psycopg2-binary==2.9.9
pywin32==306
redis==5.0.8
requests==2.32.3
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.2
